Add files via upload
Manipulating CSV Files

In this activity, you will use list comprehensions to compose a wedding invitation to send to every name on your mailing list.


Instructions


Open the file called comprehensions.py.

Run the provided program. Note that nothing forces you to write the name "properly"—e.g., as "Jane" instead of "jAnE". You will use list comprehensions to fix this.


First, use list comprehensions to create a new list that contains the lowercase version of each of the names your user provided.
Then, use list comprehensions to create a new list that contains the title-cased versions of each of the names in your lower-cased list.





Bonuses


Instead of creating a lower-cased list and then a title-cased list, create the title-cased list in a single comprehension.



Hints


See the documentation for the title method.